fix: правки замечаний
