+++ packages/client/src/store/services/chatService.ts
@@ -1,15 +1,6 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
 import { API, METHODS } from '@/utils'
-
-type GetTopicsDTO = {
-  title: string
-  author: string
-  content?: string
-}
-
-type AddTopicDTO = {
-  message: string
-}
+import { GetTopicsDTO, AddTopicDTO } from '@/types'
 
 export const chatApi = createApi({
   reducerPath: 'chatApi',
